HelpChannels: make category checks direct & efficient
Replace retrieval of all channels of a category with a direct comparison
of the categories themselves. In the case of the `on_message` listener,
the change enables the check to be done before the lock acquisition.
This is because it doesn't rely on the channels in the category to be up
to date. In fact, it doesn't even need the category object so it can
exit early without needing to wait for the cog to be ready.